test helpers: do not use default build system
Many of the executeSwift* command helpers accept a build system
provider, which was defaulted to the native build system.  Since we are
augmenting the tests to also run against the SwiftBuild, setting a
default value may lead to an undesirable test behaviour.

Update all the executeSwift* commands to no set a default build system
provider.